Rename linkBuilders -> eventLinks
