+++ lib/userEvents/eventLinks/core.js
