+++ lib/userEvents/eventLinks/getLinksForEvent.js
