+++ lib/userEvents/eventLinks/linkToViewCreateEventItem.js
