+++ lib/userEvents/eventLinks/linkToViewIssueCommentEventComment.js
