+++ lib/userEvents/eventLinks/linkToViewIssueCommentEventIssue.js
