+++ lib/userEvents/reportOnUserEvents.js
@@ -1,6 +1,6 @@
 
 const getDataForEvent = require('./eventData/getDataForEvent');
-const getLinksForEvent = require('./linkBuilders/getLinksForEvent');
+const getLinksForEvent = require('./eventLinks/getLinksForEvent');
 
 const describeTime = require('../describeTime');
